requirements.txt
torch
transformers
peft
numpy
pandas
huggingface_hub
tqdm
scikit-learn
plotly
umap-learn
